WSG v2
Introducing version 2 of my customizable wheel spacer model generator, designed for precise mock-up applications and to accommodate hubcentric mating surfaces. This tool allows users to quickly and easily generate wheel adapters tailored to specific vehicle needs, ideal for testing and visualizing wheel fitments.

The generator offers flexibility to adjust the thickness, diameter, bolt patterns, and the option to incorporate nuts and bolts for creating thicker adapters, ensuring compatibility with a wide range of vehicles. It is perfect for automotive designers, engineers, and hobbyists, providing a fast, cost-effective method for creating accurate prototypes to visualize wheel clearances and alignments.

Features:

Customizable parameters for thickness, diameter, bolt patterns, and hub options.
Accurate wheel adapter model for realistic mock-ups and fitment tests.
Compatibility with hubcentric mating surfaces for a secure and precise fit.
Designed to allow for the use of nuts and bolts as extended lugs.
Ideal for automotive engineers, designers, and enthusiasts working on wheel fitment projects.
Disclaimer: This model is intended for mock-up purposes only and must be CNC machined for use in any application, even for testing purposes. It is not suitable for 3D printing or use in actual driving conditions. If you would like a 3D printable version the thinner version 1 is suitable for mockups only and is 3D printable. It should also be noted that no tolerances have been baked into the model I recommend at least 0.24 mm be added to your measurements if you decide to send this out to be CNC machined. This model is 100% experimental and you assume all liability by downloading and manufacturing it.